Add watchCommand config option
watchCommand is specified as a list of strings. The first entry is the
application to run. The rest are arguments. Arguments don't need to be
quoted as the shell is not involved in running the command.
"{{vid}}" will be replaced with the Id of the video to watch.

The default command is still to call xdg-open with the YT URL