Use actual video publication date if availiable
This commit add usage of the video publication date in addition to the
"Added to playlist" date that was used up to this point. This improves
the video list behavior for channels which upload videos and add them to
the playlists a while before the actual publication date. Before these
published, but old in the playlist videos would show up at their upload
date instead of the publication date which is the relevant one for this
application. The "added to playlist" date is still recorded and being kept.